Make docker determine the 'last tor version' and 'last torsocks version'
This copies the logic from last_tor_version.sh and last_torsocks_version.sh
into the Dockerfile itself. The Tor version can still be specified with the
`tor_version` and `torsocks_version` arguments, as before. If these arguments
are missing, the last versions are determined by the regular expression.

This has a couple of advantages:

1) Users need not have git, make or docker-compose installed to build the image.
2) Users can build the image entirely inside a container which reduces the trust
   needed in the safety of the build instructions.
3) `git tag` is much quicker than `git ls-remote` since it operates on local data.